requirements.txt
Django>=5.0,<6
psycopg2-binary>=2.9.9
django-environ>=0.11.2
